+++ src/main/java/brig/concord/meta/model/LogYamlStepMetaType.java
@@ -1,0 +1,31 @@
+package brig.concord.meta.model;
+
+import org.jetbrains.yaml.meta.model.YamlMetaType;
+
+import java.util.Map;
+import java.util.Set;
+import java.util.function.Supplier;
+
+@SuppressWarnings("UnstableApiUsage")
+public class LogYamlStepMetaType extends IdentityMetaType {
+
+    private static final LogYamlStepMetaType INSTANCE = new LogYamlStepMetaType();
+
+    public static LogYamlStepMetaType getInstance() {
+        return INSTANCE;
+    }
+
+    private static final Map<String, Supplier<YamlMetaType>> features = Map.of(
+            "logYaml", StringMetaType::getInstance,
+            "name", StringMetaType::getInstance,
+            "meta", StepMetaMetaType::getInstance);
+
+    protected LogYamlStepMetaType() {
+        super("LogYaml", "logYaml", Set.of("logYaml"));
+    }
+
+    @Override
+    public Map<String, Supplier<YamlMetaType>> getFeatures() {
+        return features;
+    }
+}