+++ src/main/java/brig/concord/meta/model/StepElementMetaType.java
@@ -23,6 +23,7 @@ public class StepElementMetaType extends IdentityElementMetaType {
             TaskStepMetaType.getInstance(),
             CallStepMetaType.getInstance(),
             LogStepMetaType.getInstance(),
+            LogYamlStepMetaType.getInstance(),
             IfStepMetaType.getInstance(),
             CheckpointStepMetaType.getInstance(),
             SetStepMetaType.getInstance(),